+++ TimeSheet.Domain/Interfaces/IProjectDao.cs
@@ -4,6 +4,6 @@ namespace TimeSheet.Domain.Interfaces;
 
 public interface IProjectDao
 {
-    Task<int> Create(Project project);
+    Task<long> Create(Project project);
     Task<Project?> GetByTicker(string ticker);
 }
